Make hring-h2pec use maybe-uring's abstraction
...allowing it to run on macOS
