+++ test-crates/hring-h2spec/Cargo.toml
@@ -12,10 +12,11 @@ tokio = { version = "1.23.0", features = ["full"] }
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
 which = "4.4.0"
+maybe-uring = { version = "0.1.0", path = "../../crates/maybe-uring", features = ["net"] }
 
 [profile.dev.package."*"]
 opt-level = 2
 
 [features]
 default = ["tokio-uring"]
-tokio-uring = ["hring/tokio-uring"]
+tokio-uring = ["hring/tokio-uring", "maybe-uring/tokio-uring"]
